Release 1.1

+++ src/bayakm/parameters.py
@@ -1,8 +1,10 @@
+import os.path
+
 import yaml
 from baybe.parameters import SubstanceParameter, NumericalDiscreteParameter, NumericalContinuousParameter
 from baybe.utils.interval import Interval
 
-from src.bayakm.dir_paths import DirPaths
+from src.environment_variables.dir_paths import DirPaths
 
 SMILES = str
 
@@ -68,20 +70,27 @@ def build_param_list() -> list[SubstanceParameter | NumericalDiscreteParameter |
 
 def load_yaml() -> dict | None:
     dirs = DirPaths()
-    try:
-        with open(dirs.param_path, "r") as f:
-            yaml_string: str = f.read()
-    except FileNotFoundError:
-        print(f"Could not locate parameters.yaml at {dirs.param_path}.")
-        return {}
+    if os.path.exists(dirs.environ):
+        try:
+            with open(dirs.return_file_path("parameters"), "r") as f:
+                yaml_string: str = f.read()
+        except FileNotFoundError:
+            print(f"Could not locate parameters.yaml at {dirs.return_file_path("parameters")}.")
+            return {}
 
-    return yaml.safe_load(yaml_string)
+        return yaml.safe_load(yaml_string)
+    else:
+        return {}
 
 
 def save_yaml(yaml_dict: dict):
     dirs = DirPaths()
-    with open(dirs.param_path, "w") as f:
-        yaml.dump(yaml_dict, f)
+    if os.path.exists(dirs.environ):
+        try:
+            with open(dirs.return_file_path("parameters"), "w") as f:
+                yaml.dump(yaml_dict, f)
+        except FileNotFoundError:
+            print(f"Parameter file not found at {dirs.return_file_path("parameters")}")
 
 
 def write_to_parameters_file(
